add sell item route and handler

+++ internal/models/item.go
@@ -5,6 +5,6 @@ type Item struct {
 	Name        string `json:"name"`
 	Description string `json:"description"`
 	Level       uint8  `json:"level"`
-	Price       uint   `json:"price"`
+	Price       uint64 `json:"price"`
 	OwnerID     uint   `json:"owner_id"`
 }
